Tokenize sentences using 'Vibrato'

Tokenize sentences using 'Vibrato'

tokenize(
  x,
  text_field = "text",
  docid_field = "doc_id",
  sys_dic = "",
  user_dic = "",
  split = FALSE,
  mode = c("parse", "wakati")
)

Arguments

x

A data.frame like object or a character vector to be tokenized.

text_field

String or symbol; column name where to get texts to be tokenized.

docid_field

String or symbol; column name where to get identifiers of texts.

sys_dic

Character scalar; path to the system dictionary for 'Vibrato'.

user_dic

Character scalar; path to the user dictionary for 'Vibrato'.

split

Logical. If passed as `TRUE`, the function internally splits the sentence into sub-sentences using stringi::stri_split_boundaries(type = "sentence").

mode

Character scalar to switch output format.

Value

A tibble or a named list of tokens.